change field to int

+++ webapp/forms.py
@@ -83,9 +83,9 @@ class SimpleSearchForm(forms.Form):
 class OverlapForm(forms.Form):
     sequence1 = forms.CharField(label='Sequenz 1', widget=forms.Textarea(attrs={"class": "textarea-input"}))
     sequence2 = forms.CharField(label='Sequenz 2', widget=forms.Textarea(attrs={"class": "textarea-input"}))
-    match = forms.FloatField(label='Match', min_value=-1000, max_value=1000, initial=-1, widget=forms.NumberInput(attrs={'size': '10'}))
-    mismatch = forms.FloatField(label='Mismatch', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
-    gap_penalty = forms.FloatField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
+    match = forms.IntegerField(label='Match', min_value=-1000, max_value=1000, initial=-1, widget=forms.NumberInput(attrs={'size': '10'}))
+    mismatch = forms.IntegerField(label='Mismatch', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
+    gap_penalty = forms.IntegerField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
     choices = [('distance', 'Distanz'), ('similarity', 'Ähnlichkeit')]
     optimization_field = forms.ChoiceField(widget=forms.RadioSelect, choices=choices, initial='distance', label='Optimieren nach:')
 
@@ -109,9 +109,9 @@ class OverlapForm(forms.Form):
 class SmithWatermanForm(forms.Form):
     sequence1 = forms.CharField(label='Sequenz 1',  widget=forms.Textarea(attrs={"class": "textarea-input"}))
     sequence2 = forms.CharField(label='Sequenz 2', widget=forms.Textarea(attrs={"class": "textarea-input"}))
-    match = forms.FloatField(label='Match', min_value=-1000, max_value=1000, initial=1,widget=forms.NumberInput(attrs={'size': '10'}))
-    mismatch = forms.FloatField(label='Mismatch', min_value=-1000, max_value=1000, initial=-1,widget=forms.NumberInput(attrs={'size': '10'}))
-    gap_penalty = forms.FloatField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1,widget=forms.NumberInput(attrs={'size': '10'}))
+    match = forms.IntegerField(label='Match', min_value=-1000, max_value=1000, initial=1,widget=forms.NumberInput(attrs={'size': '10'}))
+    mismatch = forms.IntegerField(label='Mismatch', min_value=-1000, max_value=1000, initial=-1,widget=forms.NumberInput(attrs={'size': '10'}))
+    gap_penalty = forms.IntegerField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1,widget=forms.NumberInput(attrs={'size': '10'}))
 
     def is_valid(self):
             valid = super().is_valid()
@@ -133,9 +133,9 @@ class SmithWatermanForm(forms.Form):
 class NeedlemanWunschForm(forms.Form):
     sequence1 = forms.CharField(label='Sequenz 1', widget=forms.Textarea(attrs={"class": "textarea-input"}))
     sequence2 = forms.CharField(label='Sequenz 2',  widget=forms.Textarea(attrs={"class": "textarea-input"}))
-    match = forms.FloatField(label='Match', min_value=-1000, max_value=1000, initial=0, widget=forms.NumberInput(attrs={'size': '10'}))
-    mismatch = forms.FloatField(label='Mismatch', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
-    gap_penalty = forms.FloatField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
+    match = forms.IntegerField(label='Match', min_value=-1000, max_value=1000, initial=0, widget=forms.NumberInput(attrs={'size': '10'}))
+    mismatch = forms.IntegerField(label='Mismatch', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
+    gap_penalty = forms.IntegerField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
     choices = [('distance', 'Distanz'), ('similarity', 'Ähnlichkeit')]
     optimization_field = forms.ChoiceField(widget=forms.RadioSelect, choices=choices, initial='distance', label='Optimieren nach:')
 
@@ -161,10 +161,10 @@ class NeedlemanWunschForm(forms.Form):
 class GlocalAlignmentForm(forms.Form):
     sequence1 = forms.CharField(label='Sequenz 1', widget=forms.Textarea(attrs={"class": "textarea-input"}))
     sequence2 = forms.CharField(label='Sequenz 2', widget=forms.Textarea(attrs={"class": "textarea-input"}))
-    match = forms.FloatField(label='Match', min_value=-1000, max_value=1000, initial=-1, widget=forms.NumberInput(attrs={'size': '10'}))
-    mismatch = forms.FloatField(label='Mismatch', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
-    gap_penalty = forms.FloatField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1,widget=forms.NumberInput(attrs={'size': '10'}))
-    threshold = forms.FloatField(label='Schwellenwert', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
+    match = forms.IntegerField(label='Match', min_value=-1000, max_value=1000, initial=-1, widget=forms.NumberInput(attrs={'size': '10'}))
+    mismatch = forms.IntegerField(label='Mismatch', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
+    gap_penalty = forms.IntegerField(label='Gap-Score', min_value=-1000, max_value=1000, initial=1,widget=forms.NumberInput(attrs={'size': '10'}))
+    threshold = forms.IntegerField(label='Schwellenwert', min_value=-1000, max_value=1000, initial=1, widget=forms.NumberInput(attrs={'size': '10'}))
     choices = [('distance', 'Distanz'), ('similarity', 'Ähnlichkeit')] 
     optimization_field = forms.ChoiceField(widget=forms.RadioSelect, choices=choices, initial='distance', label='Optimieren nach:')
